Start blockquote functionality and move non components into separate folder
