+++ src/components/Fields/RichText/index.js
@@ -1,4 +1,4 @@
 export { default } from './RichText';
 export { default as Field } from './Field';
 export { default as RichTextField } from './Field';
-export { MODES } from './options';
+export { MODES } from './lib/options';
