+++ src/components/Fields/RichText/lib/serialize.js
