+++ src/components/Fields/RichText/lib/withNormalize.js
