+++ src/components/Fields/RichText/lib/withShortcuts.js
